ThermalCamera_standalone/requirements.txt
numpy~=1.21.6
imageio~=2.21.2
opencv-python~=4.5.5.64
matplotlib~=3.5.3
scikit-image~=0.19.3
pandas~=1.3.5
scipy~=1.7.3
Pillow~=9.2.0
pip~=22.2.2
pytz~=2022.2.1
setuptools~=65.3.0
openpyxl~=3.0.10
python-dateutil~=2.8.2
tifffile~=2021.11.2
Shapely~=1.8.4
packaging~=21.3
networkx~=2.6.3
six~=1.16.0
fonttools~=4.37.1
pyparsing~=3.0.9
cycler~=0.11.0
kiwisolver~=1.4.4